fix subsidy bind

+++ osznconnection/osznconnection-file-handling/src/main/java/org/complitex/osznconnection/file/service_provider/ServiceProviderAdapter.java
@@ -160,7 +160,7 @@ public class ServiceProviderAdapter extends AbstractBean {
         if (cursor.getData().size() == 1) {
             // если установлена опция перезаписи номера л/с ПУ номером л/с МН и номер л/с ПУ в файле запроса равен 0
             // и получена только одна запись из МН для данного адреса, то запись считаем связанной
-            if (updatePUAccount && 0 == Integer.valueOf(spAccountNumber)) {
+            if (updatePUAccount && 0 == Long.valueOf(spAccountNumber)) {
 
                 request.setAccountNumber(cursor.getData().get(0).getAccCode());
                 request.setStatus(RequestStatus.ACCOUNT_NUMBER_RESOLVED);
